requirements.txt
pymysql
pydantic
pydantic[email]
passlib[bcrypt]
python-dotenv
sqlalchemy
fastapi  
uvicorn
boto3
python-jose
python-multipart
requests
